+++ theme/src/components/widgets/widget-header.spec.js
@@ -8,11 +8,7 @@ const aside = <div className='sidebar-content'>Sidebar</div>
 describe('WidgetHeader', () => {
   it('matches the snapshot', () => {
     const widgetTitle = 'Neat & Interesting Widget'
-    const { asFragment } = render(
-      <WidgetHeader aside={aside} platform='GitHub'>
-        {widgetTitle}
-      </WidgetHeader>
-    )
+    const { asFragment } = render(<WidgetHeader aside={aside}>{widgetTitle}</WidgetHeader>)
     expect(asFragment()).toMatchSnapshot()
   })
 })